+++ cranko-release-info-v1
[[projects]]
qnames = ["tectonic_xdv", "cargo"]
version = "0.2.2+20250213"
age = 0

[[projects]]
qnames = ["tectonic_errors", "cargo"]
version = "0.2.1+20250213"
age = 0

[[projects]]
qnames = ["tectonic_xetex_format", "cargo"]
version = "0.3.2+20250213"
age = 0

[[projects]]
qnames = ["tectonic_status_base", "cargo"]
version = "0.2.1+20250213"
age = 0

[[projects]]
qnames = ["tectonic_io_base", "cargo"]
version = "0.4.3+20250213"
age = 0

[[projects]]
qnames = ["tectonic_geturl", "cargo"]
version = "0.3.2+20250213"
age = 0

[[projects]]
qnames = ["tectonic_docmodel", "cargo"]
version = "0.2.2+20250213"
age = 0

[[projects]]
qnames = ["tectonic_dep_support", "cargo"]
version = "0.1.1+20250213"
age = 0

[[projects]]
qnames = ["tectonic_cfg_support", "cargo"]
version = "0.1.4+20250213"
age = 0

[[projects]]
qnames = ["tectonic_bundles", "cargo"]
version = "0.3.1+20250213"
age = 0

[[projects]]
qnames = ["tectonic_bridge_icu", "cargo"]
version = "0.2.1+20250213"
age = 0

[[projects]]
qnames = ["tectonic_bridge_graphite2", "cargo"]
version = "0.2.2+20250213"
age = 0

[[projects]]
qnames = ["tectonic_bridge_harfbuzz", "cargo"]
version = "0.2.9+20250213"
age = 0

[[projects]]
qnames = ["tectonic_bridge_freetype2", "cargo"]
version = "0.2.0+20250213"
age = 0

[[projects]]
qnames = ["tectonic_bridge_flate", "cargo"]
version = "0.1.7+20250213"
age = 0

[[projects]]
qnames = ["tectonic_bridge_core", "cargo"]
version = "0.4.1+20250213"
age = 0

[[projects]]
qnames = ["tectonic_xetex_layout", "cargo"]
version = "0.2.4+20250213"
age = 0

[[projects]]
qnames = ["tectonic_pdf_io", "cargo"]
version = "0.4.1+20250213"
age = 0

[[projects]]
qnames = ["tectonic_engine_xetex", "cargo"]
version = "0.4.4+20250213"
age = 0

[[projects]]
qnames = ["tectonic_engine_xdvipdfmx", "cargo"]
version = "0.4.1+20250213"
age = 0

[[projects]]
qnames = ["tectonic_engine_spx2html", "cargo"]
version = "0.3.2+20250213"
age = 0

[[projects]]
qnames = ["tectonic_engine_bibtex", "cargo"]
version = "0.2.2+20250213"
age = 0

[[projects]]
qnames = ["tectonic", "cargo"]
version = "0.15.0+20250213"
age = 0

+++

+++ crates/engine_bibtex/src/entries.rs
@@ -6,7 +6,7 @@ pub(crate) struct EntryData {
     num_entry_ints: usize,
     num_entry_strs: usize,
     sort_key_num: usize,
-    entry_ints: Option<XBuf<i32>>,
+    entry_ints: Option<XBuf<i64>>,
     entry_strs: Option<XBuf<ASCIICode>>,
 }
 
@@ -21,11 +21,11 @@ impl EntryData {
         }
     }
 
-    pub fn ints(&self, pos: usize) -> i32 {
+    pub fn ints(&self, pos: usize) -> i64 {
         self.entry_ints.as_ref().unwrap()[pos]
     }
 
-    pub fn set_int(&mut self, pos: usize, val: i32) {
+    pub fn set_int(&mut self, pos: usize, val: i64) {
         self.entry_ints.as_mut().unwrap()[pos] = val;
     }
